requirements.txt
redis==3.2.1
paho-mqtt==1.4.0
netifaces==0.10.9